// src/models/TaskModel.ts
import type { TaskStateModel } from './TaskStateModel';

export type TaskModel = {
  id: string;
  name: string;
  duration: number;
  startDate: number;
  completeDate: number | null; //quando o timer chega ao final
  interruptDate: number | null; //quando o timer é interrompido
  type: keyof TaskStateModel['config']; // workTime, breakTime ou longBreakTime
};